+++ modules/light-clients/08-wasm/types/genesis.go
@@ -18,12 +18,12 @@ func NewGenesisState(contracts []Contract) *GenesisState {
 // ExportMetadata exports all the consensus metadata in the client store so they
 // can be included in clients genesis and imported by a ClientKeeper
 func (cs ClientState) ExportMetadata(store sdk.KVStore) []exported.GenesisMetadata {
-	payload := QueryMsg{
+	payload := queryMsg{
 		ExportMetadata: &exportMetadataMsg{},
 	}
 
 	ctx := sdk.NewContext(nil, tmproto.Header{Height: 1, Time: time.Now()}, true, nil) // context with infinite gas meter
-	result, err := wasmQuery[metadataResult](ctx, store, &cs, payload)
+	result, err := wasmQuery[exportMetadataResult](ctx, store, &cs, payload)
 	if err != nil {
 		panic(err)
 	}
