+++ modules/light-clients/08-wasm/types/msgs.go
@@ -2,8 +2,6 @@ package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/cosmos/ibc-go/v7/modules/core/exported"
 )
 
 var _ sdk.Msg = (*MsgStoreCode)(nil)
@@ -35,75 +33,3 @@ func (m MsgStoreCode) GetSigners() []sdk.AccAddress {
 	}
 	return []sdk.AccAddress{signer}
 }
-
-type instantiateMessage struct {
-	ClientState    *ClientState    `json:"client_state"`
-	ConsensusState *ConsensusState `json:"consensus_state"`
-}
-
-type QueryMsg struct {
-	Status               *statusMsg               `json:"status,omitempty"`
-	ExportMetadata       *exportMetadataMsg       `json:"export_metadata,omitempty"`
-	TimestampAtHeight    *timestampAtHeightMsg    `json:"timestamp_at_height,omitempty"`
-	VerifyClientMessage  *verifyClientMessageMsg  `json:"verify_client_message,omitempty"`
-	VerifyMembership     *verifyMembershipMsg     `json:"verify_membership,omitempty"`
-	VerifyNonMembership  *verifyNonMembershipMsg  `json:"verify_non_membership,omitempty"`
-	CheckForMisbehaviour *checkForMisbehaviourMsg `json:"check_for_misbehaviour,omitempty"`
-}
-
-type (
-	statusMsg            struct{}
-	exportMetadataMsg    struct{}
-	timestampAtHeightMsg struct {
-		Height exported.Height `json:"height"`
-	}
-	verifyClientMessageMsg struct {
-		ClientMessage *ClientMessage `json:"client_message"`
-	}
-	verifyMembershipMsg struct {
-		Height           exported.Height `json:"height"`
-		DelayTimePeriod  uint64          `json:"delay_time_period"`
-		DelayBlockPeriod uint64          `json:"delay_block_period"`
-		Proof            []byte          `json:"proof"`
-		Path             exported.Path   `json:"path"`
-		Value            []byte          `json:"value"`
-	}
-	verifyNonMembershipMsg struct {
-		Height           exported.Height `json:"height"`
-		DelayTimePeriod  uint64          `json:"delay_time_period"`
-		DelayBlockPeriod uint64          `json:"delay_block_period"`
-		Proof            []byte          `json:"proof"`
-		Path             exported.Path   `json:"path"`
-	}
-	checkForMisbehaviourMsg struct {
-		ClientMessage *ClientMessage `json:"client_message"`
-	}
-)
-
-type SudoMsg struct {
-	Initialize                    *initializeMsg                    `json:"initialize,omitempty"`
-	UpdateState                   *updateStateMsg                   `json:"update_state,omitempty"`
-	UpdateStateOnMisbehaviour     *updateStateOnMisbehaviourMsg     `json:"update_state_on_misbehaviour,omitempty"`
-	VerifyUpgradeAndUpdateState   *verifyUpgradeAndUpdateStateMsg   `json:"verify_upgrade_and_update_state,omitempty"`
-	CheckSubstituteAndUpdateState *checkSubstituteAndUpdateStateMsg `json:"check_substitute_and_update_state,omitempty"`
-}
-
-type (
-	initializeMsg struct {
-		ClientState    *ClientState    `json:"client_state"`
-		ConsensusState *ConsensusState `json:"consensus_state"`
-	}
-	updateStateMsg struct {
-		ClientMessage *ClientMessage `json:"client_message"`
-	}
-	updateStateOnMisbehaviourMsg struct {
-		ClientMessage *ClientMessage `json:"client_message"`
-	}
-	verifyUpgradeAndUpdateStateMsg struct {
-		UpgradeClientState         exported.ClientState    `json:"upgrade_client_state"`
-		UpgradeConsensusState      exported.ConsensusState `json:"upgrade_consensus_state"`
-		ProofUpgradeClient         []byte                  `json:"proof_upgrade_client"`
-		ProofUpgradeConsensusState []byte                  `json:"proof_upgrade_consensus_state"`
-	}
-	checkSubstituteAndUpdateStateMsg struct{}
-)
